+++ import_export/admin.py
@@ -732,25 +732,10 @@ class ExportMixin(BaseExportMixin, ImportExportMixinBase):
         if request.POST and "export_items" in request.POST:
             # this field is instantiated if the export is POSTed from the
             # 'action' drop down
-
-            if (
-                type(self).get_valid_export_item_pks
-                != ExportMixin.get_valid_export_item_pks
-            ):
-                warnings.warn(
-                    "The 'get_valid_export_item_pks()' method is deprecated and will "
-                    "be removed in a future release. Overwrite the ModelAdmin's "
-                    "get_queryset method to filter items instead. If you want to "
-                    "filter only the exported items, overwrite the "
-                    "get_export_queryset method.",
-                    DeprecationWarning,
-                    stacklevel=2,
-                )
-
             form.fields["export_items"] = MultipleChoiceField(
                 widget=MultipleHiddenInput,
                 required=False,
-                choices=[(pk, pk) for pk in self.get_valid_export_item_pks(request)],
+                choices=[(pk, pk) for pk in queryset.values_list("pk", flat=True)],
             )
         if form.is_valid():
             file_format = formats[int(form.cleaned_data["format"])]()
@@ -771,21 +756,6 @@ class ExportMixin(BaseExportMixin, ImportExportMixinBase):
         context = self.init_request_context_data(request, form)
         request.current_app = self.admin_site.name
         return TemplateResponse(request, [self.export_template_name], context=context)
-
-    def get_valid_export_item_pks(self, request):
-        """
-        DEPRECATED: This method is deprecated and will be removed in the future.
-        Overwrite get_queryset() or get_export_queryset() instead.
-
-        Returns a list of valid pks for export.
-        This is used to validate which objects can be exported when exports are
-        triggered from the Admin UI 'action' dropdown.
-        This can be overridden to filter returned pks for performance and/or security
-        reasons.
-        :param request: The request object.
-        :returns: a list of valid pks (by default is all pks in table).
-        """
-        return self.model.objects.all().values_list("pk", flat=True)
 
     def changelist_view(self, request, extra_context=None):
         if extra_context is None: