+++ tests/core/tests/admin_integration/test_action_export.py
@@ -130,44 +130,6 @@ class ExportActionAdminIntegrationTest(AdminTestMixin, TestCase):
             self.assertTrue(200 <= response.status_code <= 399)
             mock_export_admin_action.assert_called()
 
-    def test_export_admin_action_with_restricted_pks(self):
-        data = {
-            "format": "0",
-            "export_items": [str(self.cat1.id)],
-            **self.resource_fields_payload,
-        }
-        # mock returning a set of pks which is not in the submitted range
-        with mock.patch(
-            "import_export.admin.ExportMixin.get_valid_export_item_pks"
-        ) as mock_valid_pks:
-            mock_valid_pks.return_value = [999]
-            response = self.client.post(self.category_export_url, data)
-            self.assertEqual(response.status_code, 200)
-            self.assertIn(
-                "Select a valid choice. "
-                f"{self.cat1.id} is not one of the available choices.",
-                response.content.decode(),
-            )
-
-    def test_export_admin_action_with_restricted_pks_deprecated(self):
-        data = {
-            "format": "0",
-            "export_items": [str(self.cat1.id)],
-            **self.resource_fields_payload,
-        }
-        with mock.patch(
-            "core.admin.CategoryAdmin.get_valid_export_item_pks"
-        ) as mock_valid_pks:
-            mock_valid_pks.return_value = [999]
-            with self.assertWarnsRegex(
-                DeprecationWarning,
-                r"The 'get_valid_export_item_pks\(\)' method is deprecated and will be "
-                "removed in a future release. Overwrite the ModelAdmin's get_queryset "
-                "method to filter items instead. If you want to filter only the "
-                "exported items, overwrite the get_export_queryset method.",
-            ):
-                self.client.post(self.category_export_url, data)
-
     def _perform_export_action_calls_modeladmin_get_queryset_test(self, data):
         # Issue #1864
         # ModelAdmin's get_queryset should be used in the ModelAdmin mixins
@@ -179,6 +141,10 @@ class ExportActionAdminIntegrationTest(AdminTestMixin, TestCase):
             mock_queryset = mock.MagicMock(name="MockQuerySet")
             mock_queryset.filter.return_value = mock_queryset
             mock_queryset.order_by.return_value = mock_queryset
+            mock_queryset.values_list.return_value = [
+                str(self.cat1.id),
+                str(self.cat2.id),
+            ]
 
             mock_modeladmin_get_queryset.return_value = mock_queryset