+++ docs/changelog.rst
@@ -3,7 +3,12 @@ Changelog
 
 .. warning::
 
-    Version 4 introduces breaking changes.  Please refer to :doc:`release notes<release_notes>`.
+    Version 5 introduces breaking changes.  Please refer to :doc:`release notes<release_notes>`.
+
+5.0.0 (unreleased)
+------------------
+
+- Removed the deprecated :meth:`~import_export.admin.ExportMixin.get_valid_export_item_pks` method in favour of :meth:`~import_export.admin.ExportMixin.get_queryset`
 
 4.2.0 (unreleased)
 ------------------