+++ docs/release_notes.rst
@@ -2,6 +2,16 @@
 Release Notes
 =============
 
+v5.0 (Future release)
+----
+
+Removed the deprecated :meth:`~import_export.admin.ExportMixin.get_valid_export_item_pks` method in favour
+of :meth:`~import_export.admin.ExportMixin.get_queryset`. Use the ModelAdmin's 
+:meth:`~import_export.admin.ExportMixin.get_queryset` or :meth:`~import_export.admin.ExportMixin.get_export_queryset`
+instead.
+
+See `PR 1890 <https://github.com/django-import-export/django-import-export/issues/1890>`_.
+
 v4.2
 ----
 
@@ -11,7 +21,7 @@ This should have no impact on existing implementations.
 
 This change also made :meth:`~import_export.admin.ExportMixin.get_valid_export_item_pks` obsolete, as the
 ModelAdmin's :meth:`~import_export.admin.ExportMixin.get_queryset` or 
-:meth:`~import_export.admin.ExportMixin.get_export_queryset` can be used insted. The
+:meth:`~import_export.admin.ExportMixin.get_export_queryset` can be used instead. The
 :meth:`~import_export.admin.ExportMixin.get_valid_export_item_pks` method is now deprecated.
 
 See `PR 1890 <https://github.com/django-import-export/django-import-export/issues/1890>`_.